fn_802633B0: type the icon row index as int
s32 is signed long here, which MWCC keeps distinct from int; the row index is
an int in the target. Also reverts the previous commit's u8 player count: it
bought 24 rows but the clrlwi it emits is one the target does not have, so the
type was wrong and the gain was a register-allocation accident.

+++ src/melee/mn/mncharsel.c
@@ -3539,7 +3539,7 @@ void fn_802633B0(HSD_GObj* gobj)
     f32 new_pos;
     s32 page;
     s32 new_page;
-    s32 row_idx;
+    int row_idx;
 
     tag = gobj->user_data;
     port = tag->port;
@@ -3957,7 +3957,7 @@ s32 mnCharSel_802640A0(void)
     HSD_JObj* jobj;
     HSD_Text* text;
     s32 ctx;
-    u8 num_players;
+    s32 num_players;
     s32 row_a;
     GXColor* color_ptr;
     GXColor* color2_ptr;
